+++ app/models/guild.rb
@@ -30,6 +30,7 @@ class Guild < ApplicationRecord
   has_many :character_invitations, through: :guild_invites, source: :character
 
   has_many :banks, dependent: :destroy
+  has_many :bank_requests, -> { distinct }, through: :banks
 
   def self.cache_key(guilds)
     {