+++ spec/api/v1/bank_requests_controller_spec.rb
@@ -1,4 +1,61 @@
 RSpec.describe 'BankRequests API' do
+  describe 'GET#index' do
+    let!(:user) { create :user }
+    let!(:guild) { create :guild }
+    let!(:character) { create :character, user: user, guild: guild }
+    let!(:guild_role) { create :guild_role, guild: guild, character: character, name: 'ba' }
+    let!(:bank) { create :bank, guild: guild }
+    let!(:game_item) { create :game_item }
+    let!(:bank_cell) { create :bank_cell, bank: bank, game_item: game_item, amount: 10 }
+    let!(:bank_request1) { create :bank_request, bank: bank, character: character, game_item: game_item, requested_amount: 1 }
+    let!(:bank_request2) { create :bank_request, game_item: game_item, requested_amount: 1 }
+
+    it_behaves_like 'API auth without token'
+    it_behaves_like 'API auth with invalid token'
+    it_behaves_like 'API auth unconfirmed'
+
+    context 'with valid user token in params' do
+      let(:access_token) { JwtService.new.json_response(user: user)[:access_token] }
+
+      context 'for unexisted guild' do
+        before { get '/api/v1/bank_requests.json', params: { access_token: access_token, guild_id: 'unexisted' } }
+
+        it 'returns status 404' do
+          expect(response.status).to eq 404
+        end
+
+        it 'and returns errors' do
+          expect(JSON.parse(response.body)).to eq('error' => 'Object is not found')
+        end
+      end
+
+      context 'for existed guild' do
+        before { get '/api/v1/bank_requests.json', params: { access_token: access_token, guild_id: guild.id } }
+
+        it 'returns status 200' do
+          expect(response.status).to eq 200
+        end
+
+        it 'and returns only guild bank requests' do
+          result = JSON.parse(response.body)['bank_requests']
+
+          expect(result.size).to eq 1
+          expect(result[0]['id']).to eq bank_request1.id
+        end
+
+        %w[id character_name bank_name game_item_name requested_amount].each do |attr|
+          it "and contains bank request #{attr}" do
+            expect(response.body).to have_json_path("bank_requests/0/#{attr}")
+          end
+        end
+      end
+    end
+
+    def do_request(headers = {})
+      get '/api/v1/bank_requests.json', params: { guild_id: guild.id }, headers: headers
+    end
+  end
+
   describe 'POST#create' do
     let!(:user) { create :user }
     let!(:guild) { create :guild }