+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - render bank cells
 - BankRequest model
 - creating bank requests
+- render bank requests
 
 ## [1.0.1] - 2019-12-11
 ### Added